- throw IOException if renaming of transactional cache failed

+++ src/main/java/de/ingrid/interfaces/csw/cache/AbstractFileCache.java
@@ -413,7 +413,12 @@ public abstract class AbstractFileCache<T> implements DocumentCache<T> {
             tmpDir.delete();
             File workDir = this.getWorkPath();
             originalDir.renameTo(tmpDir);
-            workDir.renameTo(originalDir);
+            if (!originalDir.renameTo(tmpDir)) {
+                throw new IOException("Could not rename '"+originalDir+"' to '"+tmpDir+"'.");
+            }
+            if (!workDir.renameTo(originalDir)) {
+                throw new IOException("Could not rename '"+workDir+"' to '"+originalDir+"'.");
+            }
             FileUtils.deleteRecursive(tmpDir);
 
             this.inTransaction = false;
